+++ src/com/android/incallui/InCallActivity.java
@@ -882,7 +882,7 @@ public class InCallActivity extends Activity {
         ActionBar bar = getActionBar();
 
         for (int i = 0; i < phoneCount; i++) {
-            int[] subId = CallList.getInstance().getSubId(i);
+            long[] subId = CallList.getInstance().getSubId(i);
             if (subId != null && CallList.getInstance().hasAnyLiveCall(subId[0])) {
                 if (!mDsdaTabAdd[i]) {
                     addDsdaTab(i);
@@ -950,7 +950,7 @@ public class InCallActivity extends Activity {
             //setting active subscription automatically when call on one sub
             //ends and it's corresponding tab is removed.For such cases active
             //subscription will be set by InCallPresenter.attemptFinishActivity.
-            int[] subId = CallList.getInstance().getSubId(mPhoneId);
+            long[] subId = CallList.getInstance().getSubId(mPhoneId);
             if (tabCount != TAB_COUNT_ONE && CallList.getInstance().hasAnyLiveCall(subId[0])
                     && (CallList.getInstance().getActiveSubscription() != subId[0])) {
                 Log.i(this, "Switch to other active sub: " + subId[0]);
